+++ mux.go
@@ -59,4 +59,5 @@ type MuxContext interface {
 	FileAttachment(filepath, filename string) error // 将指定的文件以有效的方式写入主体流, 在客户端，文件通常会以给定的文件名下载
 	Write(p []byte) (int, error)                    // 写入响应字节流,当此方法执行完毕时应中断后续流程
 	SSE(message *SSE) (err error)                   // 发送SSE消息，只需要写入消息流即可，内部不需要处理响应头
+	FastApiContext() *Context                       // 获取内嵌的 *Context，用于框架内部获取请求上下文
 }